web/requirements.txt
autopep8==2.0.2
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
Flask==2.2.3
fuzzywuzzy==0.18.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
Levenshtein==0.20.9
MarkupSafe==2.1.2
nltk==3.8.1
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.20.9
rapidfuzz==2.13.7
regex==2023.3.23
six==1.16.0
tqdm==4.65.0
urllib3==1.26.15
Werkzeug==2.2.3
